Retry floating license (#1)
* Bump version of actions/checkout used by test workflow

Ran tests locally, passed.

* Warn about deprecated version

When the version of Advanced Installer is deprecated, this changes the behavior
from an exception to a logged warning.

* wip: Floating license support

Enables one retry to acquire a floating license slot. Improvement would be
to implement a retry loop for a specified number of seconds.

* Update action input specs

* Pass options to getExecOutput

To prevent the `exec` from failing internall, we need to pass options
so that control returns to us in order to check the return code.

* Use correct options

* Implement retry loop

* Logic error in duration

* Update README.md

* Update README.md